+++ simple/library/init.php
@@ -30,7 +30,7 @@
  * - text_domain
  *
  * @copyright  2016 WebMan Design, Oliver Juhas
- * @license    GPL-2.0+, http://www.gnu.org/licenses/gpl-2.0.html
+ * @license    GPL-3.0, http://www.gnu.org/licenses/gpl-3.0.html
  *
  * @link  https://github.com/webmandesign/webman-theme-framework
  * @link  http://www.webmandesign.eu